split hook into smaller hooks
