+++ packages/twenty-front/src/modules/ui/utilities/focus/hooks/useResetFocusStackToFocusItem.ts
@@ -1,0 +1,28 @@
+import { focusStackState } from '@/ui/utilities/focus/states/focusStackState';
+import { FocusStackItem } from '@/ui/utilities/focus/types/FocusStackItem';
+import { currentHotkeyScopeState } from '@/ui/utilities/hotkey/states/internal/currentHotkeyScopeState';
+import { previousHotkeyScopeFamilyState } from '@/ui/utilities/hotkey/states/internal/previousHotkeyScopeFamilyState';
+import { HotkeyScope } from '@/ui/utilities/hotkey/types/HotkeyScope';
+import { useRecoilCallback } from 'recoil';
+
+export const useResetFocusStackToFocusItem = () => {
+  return useRecoilCallback(
+    ({ set }) =>
+      ({
+        focusStackItem,
+        hotkeyScope,
+        memoizeKey,
+      }: {
+        focusStackItem: FocusStackItem;
+        hotkeyScope: HotkeyScope;
+        memoizeKey: string;
+      }) => {
+        set(focusStackState, [focusStackItem]);
+
+        // TODO: Remove this once we've migrated hotkey scopes to the new api
+        set(previousHotkeyScopeFamilyState(memoizeKey), null);
+        set(currentHotkeyScopeState, hotkeyScope);
+      },
+    [],
+  );
+};
